requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.4
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytube==15.0.0
pytz==2024.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
youtube-transcript-api==0.5.0
yt-dlp==2024.3.10
